More interval split reg alloc to its own module
